0x02-functions, nested loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@ int main(void)
 {
 	int c;
 	unsigned long f1 = 0, f2 = 0;
-	int sum;
+	unsigned long sum;
 
 	for (c = 0; c < 50; c++)
 	{
